Change order of if conditions so that constant precedes operator
This is a syntactic trick for C/C++ programs in which you can use an assignment statement as an expression. By putting the constant before the operator, you cant use the statement as an assignment if theres a typo and == becomes =. valid code  could still be valid should this typo occur  and few would be the wiser. To fix this, simply re-order the expression: